src/utils/config.py
"""
Logging and configuration utilities
"""
import logging
import yaml
from pathlib import Path
from typing import Dict, Any

def setup_logging(log_file: str = None) -> logging.Logger:
    """Configure logging for the application"""
    logger = logging.getLogger("urban_analytics")
    logger.setLevel(logging.INFO)
    
    formatter = logging.Formatter(
        '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    )
    
    # Console handler
    console_handler = logging.StreamHandler()
    console_handler.setFormatter(formatter)
    logger.addHandler(console_handler)
    
    # File handler
    if log_file:
        file_handler = logging.FileHandler(log_file)
        file_handler.setFormatter(formatter)
        logger.addHandler(file_handler)
    
    return logger

def load_config(config_path: str = "./config/config.yaml") -> Dict[str, Any]:
    """Load YAML configuration file"""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config

def save_config(config: Dict[str, Any], config_path: str = "./config/config.yaml") -> None:
    """Save configuration to YAML file"""
    Path(config_path).parent.mkdir(parents=True, exist_ok=True)
    with open(config_path, 'w') as f:
        yaml.dump(config, f, default_flow_style=False)

logger = setup_logging()
