requirements.txt
# Data Processing & Pipeline
pandas==2.1.3
polars==0.19.19
numpy==1.24.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Data Quality
great-expectations==0.17.12
pandera==0.16.1

# Analysis & Statistics
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0

# Visualization & Dashboard
plotly==5.17.0
dash==2.14.1
matplotlib==3.8.2
seaborn==0.13.0

# Database
duckdb==0.9.1
pyarrow==13.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
logging-loki==0.3.2

# Development
jupyter==1.0.0
ipython==8.17.2
black==23.12.0
pylint==3.0.3
